requirements.txt
Django==1.8
dj-database-url==0.3.0
dj-static==0.0.6
django-realtime==1.1
django-tastypie==0.12.2-dev
futures==2.2.0
goslate==1.4.0
gunicorn==19.3.0
psycopg2==2.6
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.9.0
static3==0.5.1
wsgiref==0.1.2
